Replace checked exceptions by unchecked exceptions
This commit replaces all checked omero exceptions with RuntimeExceptions providing the original exception as cause and an explanation as error message.
This change is necessary since omero exceptions should not be handled by the portlet directly. We try not to communicate with exceptions.

Other changes in this commit include a small refactor to the method signature of loadAnnotationsForImage