Handle OMERO sessions and generate web links (#9)
* Append session ID to avoid manual login

* Use http for image download

* Add spock framework

* Fix Test class definition

* Merge spock test classes

* Rename local variable in getImageDownloadLink(long)

* Replace System.*.print statements with Runtime exceptions

Previously checked exceptions from omero were catched and printed to the console. We try to avoid checked exceptions and turn them into unchecked exceptions to avoid an exception chaining in all applications using this library.

* Add sessionUuid to the client and repair urls

This commit adds the session Uuid to the omero client. This ID can be used as temporary password and should not be shared. We need it in order to generate omero web client links without additional login in between.

* Update Changelog

* Add JavaDoc to imageViewerLink

* Rename method to match style

* Add useless test

* Fail to implement meaningful tests

Co-authored-by: Sven Fillinger <sven.fillinger@qbic.uni-tuebingen.de>